+++ pykwalify/rule.py
@@ -4,7 +4,6 @@
 
 # python stdlib
 import logging
-import os
 import re
 
 # pykwalify imports
@@ -88,10 +87,18 @@ class Rule(object):
                     t = "map"
                     self.init_type_value(t, rule, path)
                 else:
-                    raise RuleError(u"key 'type' not found in schema rule : {}".format(path))
+                    raise RuleError(
+                        msg=u"Key 'type' not found in schema rule",
+                        error_key=u"type.missing",
+                        path=path,
+                    )
             else:
                 if not isinstance(schema["type"], str):
-                    raise RuleError(u"key 'type' in schema rule is not a string type : {}".format(path))
+                    raise RuleError(
+                        msg=u"Key 'type' in schema rule is not a string type",
+                        error_key=u"type.not_string",
+                        path=path,
+                    )
 
                 self._type = schema["type"]
 
@@ -130,9 +137,17 @@ class Rule(object):
                 func_mapping[k](v, rule, path)
             elif k.startswith("schema;"):
                 log.debug(u"Found schema tag...")
-                raise RuleError(u"Schema is only allowed on top level of schema file...")
+                raise RuleError(
+                    msg=u"Schema is only allowed on top level of schema file",
+                    error_key=u"schema.not.toplevel",
+                    path=path,
+                )
             else:
-                raise RuleError(u"Unknown key: {} found : {}".format(k, path))
+                raise RuleError(
+                    msg=u"Unknown key: {} found".format(k),
+                    error_key=u"key.unknown",
+                    path=path,
+                )
 
         self.check_conflicts(schema, rule, path)
 
@@ -140,7 +155,11 @@ class Rule(object):
         """
         """
         if not isinstance(v, str):
-            raise RuleError(u"value for func keyword must be a string : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Value: {} for func keyword must be a string".format(v),
+                error_key=u"func.notstring",
+                path=path,
+            )
 
         self._func = v
 
@@ -148,7 +167,11 @@ class Rule(object):
         """
         """
         if not isinstance(v, list):
-            raise RuleError(u"Extension defenition should be a list : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Extension defenition should be a list",
+                error_key=u"extension.not_list",
+                path=path,
+            )
 
         # TODO: Add limitation that this keyword can only be used at the top level of the file
 
@@ -162,7 +185,11 @@ class Rule(object):
         allowed = ["any"]
         # ["none", "one", "all"] Is currently awaiting proper implementation
         if v not in allowed:
-            raise RuleError(u"Specefied rule in key : {} is not part of allowed rule set : {}".format(v, allowed))
+            raise RuleError(
+                msg=u"Specified rule in key: {} is not part of allowed rule set : {}".format(v, allowed),
+                error_key=u"matching_rule.not_allowed",
+                path=path,
+            )
         else:
             self._matching_rule = v
 
@@ -183,7 +210,11 @@ class Rule(object):
         self._type_class = type_class(v)
 
         if not is_builtin_type(self._type):
-            raise RuleError(u"type.unknown : {} : {}".format(self._type, path))
+            raise RuleError(
+                msg=u"Type: {} is not any of the known types".format(self._type),
+                error_key=u"type.unknown",
+                path=path,
+            )
 
     def init_matching(self, v, rule, path):
         log.debug(u"Init matching rule : {}".format(path))
@@ -191,7 +222,11 @@ class Rule(object):
         valid_values = ["any", "all", "*"]
 
         if str(v) not in valid_values:
-            raise RuleError(u"matching value: {} is not one of {}".format(str(v), valid_values))
+            raise RuleError(
+                msg=u"matching value: {} is not one of {}".format(str(v), valid_values),
+                error_key=u"matching_rule.invalid",
+                path=path,
+            )
 
         self._matching = str(v)
 
@@ -209,44 +244,76 @@ class Rule(object):
         log.debug(u"Init required value : {}".format(path))
 
         if not isinstance(v, bool):
-            raise RuleError(u"required.notbool : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Value: '{}' for required keyword must be a boolean".format(v),
+                error_key=u"required.not_bool",
+                path=path,
+            )
         self._required = v
 
     def init_pattern_value(self, v, rule, path):
         log.debug(u"Init pattern value : {}".format(path))
 
         if not isinstance(v, str):
-            raise RuleError(u"pattern.notstr : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Value of pattern keyword: '{}' is not a string".format(v),
+                error_key=u"pattern.not_string",
+                path=path,
+            )
 
         self._pattern = v
 
         if self._schema_str["type"] == "map":
-            raise RuleError(u"map.pattern : pattern not allowed inside map : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Keyword pattern is not allowed inside map",
+                error_key=u"pattern.not_allowed_in_map",
+                path=path,
+            )
 
         # TODO: Some form of validation of the regexp? it exists in the source
 
         try:
             self._pattern_regexp = re.compile(self._pattern)
         except Exception:
-            raise RuleError(u"pattern.syntaxerr : {} --> {} : {}".format(self._pattern_regexp, self._pattern_regexp, path))
+            raise RuleError(
+                msg=u"Syntax error when compiling regex pattern: {}".format(self._pattern_regexp),
+                error_key=u"pattern.syntax_error",
+                path=path,
+            )
 
     def init_enum_value(self, v, rule, path):
         log.debug(u"Init enum value : {}".format(path))
 
         if not isinstance(v, list):
-            raise RuleError(u"enum.notseq")
+            raise RuleError(
+                msg=u"Enum is not a sequence",
+                error_key=u"enum.not_seq",
+                path=path,
+            )
         self._enum = v
 
         if is_collection_type(self._type):
-            raise RuleError(u"enum.notscalar")
+            raise RuleError(
+                msg=u"Enum is not a scalar",
+                error_key=u"enum.not_scalar",
+                path=path,
+            )
 
         lookup = set()
         for item in v:
             if not isinstance(item, self._type_class):
-                raise RuleError(u"enum.type.unmatch : {} --> {} : {}".format(item, self._type_class, path))
+                raise RuleError(
+                    msg=u"Item: '{}' in enum is not of correct class type: '{}'".format(item, self._type_class),
+                    error_key=u"enum.type.unmatch",
+                    path=path,
+                )
 
             if item in lookup:
-                raise RuleError(u"enum.duplicate : {} : {}".format(item, path))
+                raise RuleError(
+                    msg=u"Duplicate items: '{}' found in enum".format(item),
+                    error_key=u"enum.duplicate_items",
+                    path=path,
+                )
 
             lookup.add(item)
 
@@ -254,31 +321,64 @@ class Rule(object):
         log.debug(u"Init assert value : {}".format(path))
 
         if not isinstance(v, str):
-            raise RuleError(u"assert.notstr : {}".format(path))
+            raise RuleError(
+                msg=u"Value: '{}' for keyword 'assert' is not a string".format(v),
+                error_key=u"assert.not_str",
+                path=path,
+            )
 
         self._assert = v
 
-        raise RuleError(u"assert.NYI-Error : {}".format(path))
+        raise RuleError(
+            msg=u"Keyword assert is not yet implemented",
+            error_key=u"assert.NotYetImplemented",
+            path=path,
+        )
 
     def init_range_value(self, v, rule, path):
         log.debug(u"Init range value : {}".format(path))
 
-        if not isinstance(v, dict):
-            raise RuleError(u"range.notmap : {} : {}".format(v, path))
-        if self._type not in ["str", "int", "map", "seq"]:
-            raise RuleError(u"range.not-supported-type : {} : {}".format(self._type, path))
+        supported_types = ["str", "int", "map", "seq"]
 
-        self._range = v  # dict that should contain min, max, min-ex, max-ex keys
+        if not isinstance(v, dict):
+            raise RuleError(
+                msg=u"Range value is not a dict type: '{}'".format(v),
+                error_key=u"range.not_map",
+                path=path,
+            )
+
+        if self._type not in supported_types:
+            raise RuleError(
+                msg=u"Range value type: '{}' is not a supported type".format(self._type),
+                error_key=u"range.not_supported_type",
+                path=path,
+            )
+
+        # dict that should contain min, max, min-ex, max-ex keys
+        self._range = v
 
         # This should validate that only min, max, min-ex, max-ex exists in the dict
         for k, v in self._range.items():
             if k not in ["max", "min", "max-ex", "min-ex"]:
-                raise RuleError(u"range.undefined key : {} : {}".format(k, path))
+                raise RuleError(
+                    msg=u"Unknown key: '{}' found in range keyword".format(k),
+                    error_key=u"range.unknown_key",
+                    path=path,
+                )
 
         if "max" in self._range and "max-ex" in self._range:
-            raise RuleError(u"range.twomax : {}".format(path))
+            raise RuleError(
+                msg=u"'max' and 'max-ex' can't be used in the same range rule",
+                error_key=u"range.max_duplicate_keywords",
+                path=path,
+            )
+
         if "min" in self._range and "min-ex" in self._range:
-            raise RuleError(u"range.twomin : {}".format(path))
+            raise RuleError(
+                msg=u"'min' and 'min-ex' can't be used in the same range rule",
+                error_key=u"range.min_duplicate_keywords",
+                path=path,
+            )
 
         max = self._range.get("max", None)
         min = self._range.get("min", None)
@@ -286,67 +386,137 @@ class Rule(object):
         min_ex = self._range.get("min-ex", None)
 
         if max is not None and not is_int(max) or is_bool(max):
-            raise RuleError(u"range.max.notint : {} : {}".format(max, path))
+            raise RuleError(
+                msg=u"Value: '{}' for 'max' keyword is not a integer".format(v),
+                error_key=u"range.max.not_int",
+                path=path,
+            )
 
         if min is not None and not is_int(min) or is_bool(min):
-            raise RuleError(u"range.min.notint : {} : {}".format(min, path))
+            raise RuleError(
+                msg=u"Value: '{}' for 'min' keyword is not a integer".format(v),
+                error_key=u"range.min.not_int",
+                path=path,
+            )
 
         if max_ex is not None and not is_int(max_ex) or is_bool(max_ex):
-            raise RuleError(u"range.max_ex.notint : {} : {}".format(max_ex, path))
+            raise RuleError(
+                msg=u"Value: '{}' for 'max-ex' keyword is not a integer".format(v),
+                error_key=u"range.max_ex.not_int",
+                path=path,
+            )
 
         if min_ex is not None and not is_int(min_ex) or is_bool(min_ex):
-            raise RuleError(u"range.min_ex.notint : {} : {}".format(min_ex, path))
+            raise RuleError(
+                msg=u"Value: '{}' for 'min-ex' keyword is not a integer".format(v),
+                error_key=u"range.min_ex.not_int",
+                path=path,
+            )
 
         if max is not None:
             if min is not None and max < min:
-                raise RuleError(u"range.maxltmin : {} < {} : {}".format(max, min, path))
+                raise RuleError(
+                    msg=u"Value for 'max' can't be less then value for 'min'. {} < {}".format(max, min),
+                    error_key=u"range.max_lt_min",
+                    path=path,
+                )
             elif min_ex is not None and max <= min_ex:
-                raise RuleError(u"range.maxleminex : {} <= {} : {}".format(max, min_ex, path))
+                raise RuleError(
+                    msg=u"Value for 'max' can't be less then value for 'min-ex'. {} <= {}".format(max, min_ex),
+                    error_key=u"range.max_le_min-ex",
+                    path=path,
+                )
         elif max_ex is not None:
             if min is not None and max_ex < min:
-                raise RuleError(u"range.maxexlemiin : {} < {} : {}".format(max_ex, min, path))
+                raise RuleError(
+                    msg=u"Value for 'max-ex' can't be less then value for 'min'. {} < {}".format(max_ex, min),
+                    error_key=u"range.max-ex_le_min",
+                    path=path,
+                )
             elif min_ex is not None and max_ex <= min_ex:
-                raise RuleError(u"range.maxexleminex : {} <= {} : {}".format(max_ex, min_ex, path))
+                raise RuleError(
+                    msg=u"Value for 'max-ex' can't be less then value for 'min-ex'. {} <= {}".format(max_ex, min_ex),
+                    error_key=u"range.max-ex_le_min-ex",
+                    path=path,
+                )
 
     def init_ident_value(self, v, rule, path):
         log.debug(u"Init ident value : {}".format(path))
 
         if v is None or isinstance(v, bool):
-            raise RuleError(u"ident.notbool : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Value: '{}' of 'ident' is not a boolean value".format(v),
+                error_key=u"ident.not_bool",
+                path=path,
+            )
 
         self._ident = bool(v)
         self._required = True
 
         if is_collection_type(self._type):
-            raise RuleError(u"ident.notscalar : {} : {}".format(self._type, path))
+            raise RuleError(
+                msg=u"Value: '{}' of 'ident' is not a scalar value".format(v),
+                error_key=u"ident.not_scalar",
+                path=path,
+            )
+
         if path == "":
-            raise RuleError(u"ident.onroot")
+            raise RuleError(
+                msg=u"Keyword 'ident' can't be on root level of schema",
+                error_key=u"ident.not_on_root_level",
+                path=path,
+            )
+
         if self._parent is None or not self._parent._type == "map":
-            raise RuleError(u"ident.notmap : {}".format(path))
+            raise RuleError(
+                msg=u"Keword 'ident' can't be inside 'map'",
+                error_key=u"ident.not_in_map",
+                path=path,
+            )
 
     def init_unique_value(self, v, rule, path):
         log.debug(u"Init unique value : {}".format(path))
 
         if not isinstance(v, bool):
-            raise RuleError(u"unique.notbool : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Value: '{}' for 'unique' keyword is not boolean".format(v),
+                error_key=u"unique.not_bool",
+                path=path,
+            )
 
         self._unique = v
 
         if is_collection_type(self._type):
-            raise RuleError(u"unique.notscalar : {} : {}".format(self._type, path))
+            raise RuleError(
+                msg=u"Type of the value: '{}' for 'unique' keyword is not a scalar type".format(self._type),
+                error_key=u"unique.not_scalar",
+                path=path,
+            )
         if path == "":
-            raise RuleError(u"unique.onroot")
+            raise RuleError(
+                msg=u"Keyword 'unique' can't be on root level of schema",
+                error_key=u"unique.not_on_root_level",
+                path=path,
+            )
 
     def init_sequence_value(self, v, rule, path):
         log.debug(u"Init sequence value : {}".format(path))
 
         if v is not None and not isinstance(v, list):
-            raise RuleError(u"sequence.notseq : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Sequence keyword is not a list",
+                error_key=u"sequence.not_seq",
+                path=path,
+            )
 
         self._sequence = v
 
         if self._sequence is None or len(self._sequence) == 0:
-            raise RuleError(u"sequence.noelem : {} : {}".format(self._sequence, path))
+            raise RuleError(
+                msg=u"Sequence contains 0 elements",
+                error_key=u"sequence.no_elements",
+                path=path,
+            )
 
         tmp_seq = []
 
@@ -365,15 +535,27 @@ class Rule(object):
     def init_mapping_value(self, v, rule, path):
         # Check for duplicate use of 'map' and 'mapping'
         if self._mapping:
-            raise RuleError(u"mapping.multiple-use : {}".format(path))
+            raise RuleError(
+                msg=u"Keywords 'map' and 'mapping' can't be used on the same level",
+                error_key=u"mapping.duplicate_keywords",
+                path=path,
+            )
 
         log.debug(u"Init mapping value : {}".format(path))
 
         if v is not None and not isinstance(v, dict):
-            raise RuleError(u"mapping.notmap : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Value for keyword 'map/mapping' is not a dict",
+                error_key=u"mapping.not_dict",
+                path=path,
+            )
 
         if v is None or len(v) == 0:
-            raise RuleError(u"mapping.noelem : {} : {}".format(v, path))
+            raise RuleError(
+                msg=u"Mapping do not contain any elements",
+                error_key=u"mapping.no_elements",
+                path=path,
+            )
 
         self._mapping = {}
         self._regex_mappings = []
@@ -387,13 +569,22 @@ class Rule(object):
                 log.debug(u"Found regex map rule")
                 regex = k.split(";", 1)
                 if len(regex) != 2:
-                    raise RuleError(u"Malformed regex key : {}".format(k))
+                    raise RuleError(
+                        msg=u"Value: '{}' for keyword regex is malformed".format(k),
+                        error_key=u"mapping.regex.malformed",
+                        path=path,
+                    )
                 else:
                     regex = regex[1]
                     try:
                         re.compile(regex)
                     except Exception as e:
-                        raise RuleError(u"Unable to compile regex '{}' '{}'".format(regex, e))
+                        log.debug(e)
+                        raise RuleError(
+                            msg=u"Unable to compile regex '{}'".format(regex),
+                            error_key=u"mapping.regex.compile_error",
+                            path=path,
+                        )
 
                     regex_rule = Rule(None, self)
                     regex_rule.init(v, u"{}/mapping;regex/{}".format(path, regex[1:-1]))
@@ -412,13 +603,25 @@ class Rule(object):
         self._default = v
 
         if is_collection_type(self._type):
-            raise RuleError(u"default.notscalar : {} : {} : {}".format(rule, path, v))
+            raise RuleError(
+                msg=u"Value: {} for keyword 'default' is not a scalar type".format(v),
+                error_key=u"default.not_scalar",
+                path=path,
+            )
 
         if self._type == "map" or self._type == "seq":
-            raise RuleError(u"default.notscalar : {} : {} : {}".format(rule, os.path.dirname(path), v))
+            raise RuleError(
+                msg=u"Value: {} for keyword 'default' is not a scalar type".format(v),
+                error_key=u"default.not_scalar",
+                path=path,
+            )
 
         if not isinstance(v, self._type_class):
-            raise RuleError(u"default.type.unmatch : {} --> {} : {}".format(v, self._type_class, path))
+            raise RuleError(
+                msg=u"Types do not match: '{}' --> '{}'".format(v, self._type_class),
+                error_key=u"default.type.unmatch",
+                path=path,
+            )
 
     def check_conflicts(self, schema, rule, path):
         log.debug(u"Checking for conflicts : {}".format(path))